requirements.txt
django>=4.0.4,<4.1.0
djangorestframework>=3.13.1,<3.14.0
flake8>=4.0.1,<4.1.0
pandas>=1.4.2,<1.5.0
Pillow>=9.1.0,<9.2.0
psycopg2>=2.9.3,<2.10.0
scrapy>=2.6.1,<2.7.0
scrapy-djangoitem>=1.1.1,<1.2.0
selenium>=4.1.5, <4.2.0
SQLAlchemy>=1.4.36,<1.5.0
webdriver-manager>=3.5.4,<3.6.0